feat(dicom): Enhance DICOM input handling with new loading functions and exceptions
