+++ src/imgtools/dicom/__init__.py
@@ -1,3 +1,10 @@
+from .input import (
+    DicomInput,
+    load_dicom,
+    load_rtstruct_dcm,
+    load_seg_dcm,
+    path_from_pathlike,
+)
 from .utils import find_dicoms, lookup_tag, similar_tags, tag_exists
 
 __all__ = [
@@ -5,4 +12,10 @@ __all__ = [
     "lookup_tag",
     "similar_tags",
     "tag_exists",
+    # input
+    "DicomInput",
+    "load_dicom",
+    "path_from_pathlike",
+    "load_rtstruct_dcm",
+    "load_seg_dcm",
 ]
