+++ src/imgtools/dicom/input/__init__.py
@@ -1,1 +1,15 @@
-from .dicom_reader import load_dicom, path_from_pathlike
+from .dicom_reader import (
+    DicomInput,
+    load_dicom,
+    load_rtstruct_dcm,
+    load_seg_dcm,
+    path_from_pathlike,
+)
+
+__all__ = [
+    "DicomInput",
+    "load_dicom",
+    "path_from_pathlike",
+    "load_rtstruct_dcm",
+    "load_seg_dcm",
+]
