+++ src/imgtools/exceptions.py
@@ -2,12 +2,22 @@ class DirectoryNotFoundError(Exception):
     pass
 
 
+####################################################################################################
+# Dicom spcific exceptions
+
+
 class InvalidDicomError(Exception):
     pass
 
 
 class NotRTSTRUCTError(InvalidDicomError):
-    """Exception raised for errors in the RTSTRUCT DICOM file."""
+    """Exception raised when loading a non-RTSTRUCT DICOM file."""
+
+    pass
+
+
+class NotSEGError(InvalidDicomError):
+    """Exception raised when loading a non-SEG DICOM file."""
 
     pass
 
@@ -16,6 +26,7 @@ class RTSTRUCTAttributeError(InvalidDicomError):
     """Exception raised for attribute errors in the RTSTRUCT DICOM file."""
 
     pass
+
 
 class MissingROIError(KeyError):
     """Custom exception for missing ROI in the structure set."""
